+++ internal/pkg/configure/tftp.go
@@ -58,6 +58,10 @@ func TFTP() (err error) {
 		}
 	}
 
+	if err := buildHostOverlays("tftp", controller.TFTP.Overlays); err != nil {
+		wwlog.Warn("host overlay could not be built: %s", err)
+	}
+
 	if !controller.TFTP.Enabled() {
 		wwlog.Warn("Warewulf does not auto start TFTP services due to disable by warewulf.conf")
 		return nil